[#216] Monitor: Use uint32_t for time values
uint16_t milliseconds counter may wrap on long operations.

Affected functions:
getTimeSec();
getTotalBalanceTimeSec();
getTotalChargeDischargeTimeSec();
getETATime();

+++ src/core/screens/ScreenCycle.cpp
@@ -36,7 +36,7 @@ namespace Screen { namespace Cycle {
 
     //TODO_NJ for cyclehistory
     uint16_t cyclesHistoryCapacity[10];
-    uint16_t cyclesHistoryTime[10];
+    uint32_t cyclesHistoryTime[10];
 
 } // namespace Screen
 } // namespace Cycle
