+++ src/core/strategy/Monitor.h
@@ -24,7 +24,7 @@
 #define MONITOR_EXTERNAL_ERROR_BATTERY_DISCONNECTED         1
 
 namespace Monitor {
-    extern uint16_t etaDeltaSec;
+    extern uint32_t etaDeltaSec;
     extern bool isBalancePortConnected;
     extern volatile uint8_t i_externalError;
 
@@ -34,10 +34,10 @@ namespace Monitor {
     void powerOff();
 
     uint32_t getTimeSec();
-    uint16_t getTotalBalanceTimeSec();
-    uint16_t getTotalChargeDischargeTimeSec();
+    uint32_t getTotalBalanceTimeSec();
+    uint32_t getTotalChargeDischargeTimeSec();
     uint16_t getTotalChargeDischargeTimeMin();
-    uint16_t getETATime();
+    uint32_t getETATime();
 
     uint8_t getChargeProcent();
